Backend-main/massuh-datalake-main/apresentame-api/requirements.txt
pandas
google-cloud
google-auth
requests
pandas_gbq
os